+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Add Symfony 3.0 support
 - Add `--no-outdated` and `--no-upgradable` flags
+- Add none-zero status codes for CI
 
 ## 0.7.0 (released 2015-11-28)